+++ hlnsrv-script.bash
@@ -1142,7 +1142,7 @@ script_install_prefix() {
 			env WINEARCH=$WINE_ARCH WINEDEBUG=-all WINEDLLOVERRIDES="mscoree=d" WINEPREFIX=$SRV_DIR wineboot --init /nogui
 			env WINEARCH=$WINE_ARCH WINEDEBUG=-all WINEPREFIX=$SRV_DIR winetricks corefonts
 			env DISPLAY=:5.0 WINEARCH=$WINE_ARCH WINEDEBUG=-all WINEPREFIX=$SRV_DIR winetricks -q vcrun2015
-			env DISPLAY=:5.0 WINEARCH=$WINE_ARCH WINEDEBUG=-all WINEPREFIX=$SRV_DIR winetricks -q dotnet472
+			env DISPLAY=:5.0 WINEARCH=$WINE_ARCH WINEDEBUG=-all WINEPREFIX=$SRV_DIR winetricks -q dotnet48
 			pkill -f Xvfb
 			if [[ "$REINSTALL_PREFIX_KEEP_DATA" =~ ^([yY][eE][sS]|[yY])$ ]]; then
 				mkdir -p "$SRV_DIR/$WINE_PREFIX_GAME_DIR"
@@ -1371,6 +1371,12 @@ script_diagnostics() {
 		echo "Notification sending service present: No"
 	fi
 	
+	if [ -f "$SRV_DIR/$WINE_PREFIX_GAME_DIR/$WINE_PREFIX_GAME_EXE" ]; then
+		echo "Game executable present: Yes"
+	else
+		echo "Game executable present: No"
+	fi
+	
 	echo "Diagnostics complete."
 }
 
@@ -1399,7 +1405,11 @@ script_install_packages() {
 			
 			if [[ "$UBUNTU_CODENAME" == "bionic" || "$UBUNTU_CODENAME" == "eoan" || "$UBUNTU_CODENAME" == "focal" ]]; then
 				#Add i386 architecture support
+				apt install --yes sudo gnupg
 				sudo dpkg --add-architecture i386
+				
+				#Install software properties common
+				sudo apt install --yes software-properties-common
 				
 				#Check codename and install config for installation
 				if [[ "$UBUNTU_CODENAME" == "bionic" ]]; then
@@ -1444,6 +1454,65 @@ script_install_packages() {
 				echo "Exiting"
 				exit 1
 			fi
+		elif [[ "$DISTRO" == "debian" ]]; then
+			#Debian distro
+			
+			#Get codename
+			DEBIAN_CODENAME=$(cat /etc/os-release | grep "^VERSION_CODENAME=" | cut -d = -f2)
+			
+			if [[ "$DEBIAN_CODENAME" == "buster" ]]; then
+				#Add i386 architecture support
+				apt install --yes sudo gnupg
+				sudo dpkg --add-architecture i386
+				
+				#Install software properties common
+				sudo apt install --yes software-properties-common
+				
+				#Add non-free repo for steamcmd
+				sudo apt-add-repository non-free
+				
+				#Check codename and install backport repo if needed
+				if [[ "$DEBIAN_CODENAME" == "buster" ]]; then
+					sudo apt-add-repository "deb http://deb.debian.org/debian $DEBIAN_CODENAME-backports main"
+					sudo apt update
+					sudo apt -t buster-backports install --yes "tmux"
+				fi
+			
+				#Add wine and libfaudio0 repositroy and install packages
+				wget -nc https://dl.winehq.org/wine-builds/winehq.key
+				sudo apt-key add winehq.key
+				
+				wget -nc https://download.opensuse.org/repositories/Emulators:/Wine:/Debian/Debian_10/Release.key
+				sudo apt-key add Release.key
+				
+				sudo apt-add-repository "deb https://dl.winehq.org/wine-builds/debian/ $DEBIAN_CODENAME main"
+				sudo apt-add-repository "deb https://download.opensuse.org/repositories/Emulators:/Wine:/Debian/Debian_10 ./"
+				
+				#Check for updates and update local repo database
+				sudo apt update
+				
+				#Install packages and enable services
+				sudo apt install --yes libfaudio0:i386
+				sudo apt install --yes libfaudio0
+				sudo apt install --yes --install-recommends winehq-staging
+				sudo apt install --yes --install-recommends steamcmd
+				sudo apt install --yes rsync cabextract unzip p7zip wget curl postfix zip jq xvfb samba winbind
+				sudo systemctl enable smbd nmbd winbind
+				sudo systemctl start smbd nmbd winbind
+				
+				#Install winetricks
+				wget https://raw.githubusercontent.com/Winetricks/winetricks/master/src/winetricks
+				sudo mv winetricks /usr/local/bin/
+				sudo chmod +x /usr/local/bin/winetricks
+			else
+				echo "Error: This version of Debian is not supported. Supported versions are: Debian 10 (Buster)"
+				echo "Exiting"
+				exit 1
+			fi
+		else
+			echo "Error: This distro is not supported. This script currently supports Arch Linux, Ubuntu 18.04 LTS (Bionic Beaver), Ubuntu 19.10 (Disco Dingo), Ubuntu 20.04 LTS (Focal Fossa), Debian 10 (Buster). If you want to try the script on your distro, install the packages manually. Check the readme for required package versions."
+			echo "Exiting"
+			exit 1
 		fi
 		
 		if [[ "$DISTRO" == "arch" ]]; then
@@ -1681,11 +1750,6 @@ script_install() {
 		DISCORD_CRASH="0"
 	fi
 	
-	echo "Enabling linger"
-	sudo mkdir -p /var/lib/systemd/linger/
-	sudo touch /var/lib/systemd/linger/$USER
-	sudo mkdir -p /home/$USER/.config/systemd/user
-	
 	echo "Installing bash profile"
 	cat > /home/$USER/.bash_profile <<- 'EOF'
 	#
@@ -1704,6 +1768,16 @@ script_install() {
 	script_install_services
 	
 	sudo chown -R $USER:users /home/$USER/.config/systemd/user
+	
+	echo "Enabling linger"
+	
+	sudo loginctl enable-linger $USER
+	
+	if [ -d /var/lib/systemd/linger/$USER ]; then
+		sudo mkdir -p /var/lib/systemd/linger/
+		sudo touch /var/lib/systemd/linger/$USER
+		sudo mkdir -p /home/$USER/.config/systemd/user
+	fi
 	
 	echo "Enabling services"
 		
@@ -1759,7 +1833,7 @@ script_install() {
 	env WINEARCH=$WINE_ARCH WINEDEBUG=-all WINEDLLOVERRIDES="mscoree=d" WINEPREFIX=$SRV_DIR wineboot --init /nogui
 	env WINEARCH=$WINE_ARCH WINEDEBUG=-all WINEPREFIX=$SRV_DIR winetricks corefonts
 	env DISPLAY=:5.0 WINEARCH=$WINE_ARCH WINEDEBUG=-all WINEPREFIX=$SRV_DIR winetricks -q vcrun2015
-	env DISPLAY=:5.0 WINEARCH=$WINE_ARCH WINEDEBUG=-all WINEPREFIX=$SRV_DIR winetricks -q dotnet472
+	env DISPLAY=:5.0 WINEARCH=$WINE_ARCH WINEDEBUG=-all WINEPREFIX=$SRV_DIR winetricks -q dotnet48
 	pkill -f Xvfb
 	EOF